+++ apps/nozomi/src/__init__.py
@@ -1,0 +1,1 @@
+"""nozomi source package."""
