+++ apps/nozomi/src/schemas/__init__.py
@@ -1,0 +1,1 @@
+"""Schema modules for nozomi."""
